refactor: extract service
Extract `AddUser` service
