+++ src/services/fetch-packument.ts
@@ -5,7 +5,7 @@ import { assertIsHttpError } from "../utils/error-type-guards";
 import { UnityPackument } from "../domain/packument";
 import { HttpErrorBase } from "npm-registry-fetch";
 import { DomainName } from "../domain/domain-name";
-import { Registry } from "../npm-client";
+import { Registry } from "./add-user-service";
 
 /**
  * Service for fetching packuments from a npm-registry.
